Retain only library parts of realn.c; canonicalise whitespace
Samtools's bam_md.c contains both library functions and the calmd
command; remove the non-library-function parts and rename to realn.c.

Detab kprobaln.c and kprobaln.h.
